refactor(acceptance-tests): reorganize setup helpers and logging functions
