+++ tests/acceptance/support/helpers/setup/build.helpers.ts
@@ -1,0 +1,15 @@
+import { spawnSync } from "node:child_process";
+
+import type { SpawnOptions } from "node:child_process";
+
+export function buildAppForAcceptanceTests(): void {
+  const spawnOptions: SpawnOptions = {
+    shell: true,
+    stdio: "inherit",
+  };
+  const buildProcess = spawnSync("pnpm run build", spawnOptions);
+  if (buildProcess.status !== 0) {
+    const errorDetails = buildProcess.error?.message ?? "Unknown error";
+    throw new Error(`Failed to build the application for acceptance tests. ${errorDetails}`);
+  }
+}
